Snowfix on header component, refactored functions declaration to arrow functions, renamed some folders
