+++ frontend/src/App.tsx
@@ -5,11 +5,11 @@ import ProductPage from "./pages/ProductPage";
 import ContactsPage from "./pages/ContactsPage";
 import BasketPage from "./pages/BasketPage";
 import OrderPage from "./pages/OrderPage";
-import Header from "./components/UserInterface/Header";
-import Footer from "./components/UserInterface/Footer";
-import MainContainer from "./components/UserInterface/MainContainer";
+import Header from "./components/ui/Header";
+import Footer from "./components/ui/Footer";
+import MainContainer from "./components/ui/MainContainer";
 
-function App() {
+const App = () => {
     return (
         <Router>
             <Header />
