--- frontend/src/components/UserInterface/Footer.tsx
@@ -1,13 +1,0 @@
-import BottomFooter from '../footer/BottomFooter';
-import SubscribeOnEmailSection from '../footer/SubscribeOnEmailSection';
-
-const Footer = () => {
-    return(
-        <footer>
-            <SubscribeOnEmailSection />
-            <BottomFooter />
-        </footer>
-    );
-}
-
-export default Footer;
