+++ frontend/src/components/ui/Footer.tsx
@@ -1,0 +1,13 @@
+import BottomFooter from "../footer/BottomFooter";
+import SubscribeOnEmailSection from "../footer/SubscribeOnEmailSection";
+
+const Footer = () => {
+    return (
+        <footer>
+            <SubscribeOnEmailSection />
+            <BottomFooter />
+        </footer>
+    );
+};
+
+export default Footer;
