+++ frontend/src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
-function MainPage() {
+const MainPage = () => {
     return <div></div>;
-}
+};
 
 export default MainPage;
